requirements.txt
blinker==1.8.2
click==8.1.7
et_xmlfile==2.0.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pycountry==24.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
